Test: Split path grammar tests into paths.test.js
